bug fixed for double

+++ src/main/java/com/alibaba/fastjson/JSONObject.java
@@ -247,7 +247,7 @@ public class JSONObject extends JSON implements Map<String, Object>, JSONAware, 
             return 0D;
         }
 
-        return castToDouble(value).floatValue();
+        return castToDouble(value).doubleValue();
     }
 
     public BigDecimal getBigDecimal(String key) {
